refac:  move BoundaryController and CollisionController from the services to the controller folder/namespace
