+++ BattleStars/Application/Controllers/BoundaryController.cs
@@ -1,6 +1,6 @@
 using BattleStars.Domain.Interfaces;
 using BattleStars.Core.Guards;
-namespace BattleStars.Application.Services;
+namespace BattleStars.Application.Controllers;
 
 /// <summary>
 /// Implements boundary enforcement in the game.
